move helper/config to bundle with the other template definitions
